organize project structure and update models location
